Update test.js
innerText retrieves and sets the content of the tag as plain text.
innerHTML retrieves and sets the same content in HTML format.
The content of all elements in the node, including script and style elements. It’s aware of formatting like spacing and line breaks and will return those, innerText is very similar to textContent, however, there are important differences between them! Put simply, innerText is aware of the rendered appearance of text, while textContent is not.